backend/requirements.txt
fastapi==0.115.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
pydantic-settings==2.7.1
apscheduler==3.10.4
python-dotenv==1.0.1
